Refactoring. Working on clean container structure + orchestration + MLflow registry
