+++ api/auth.py
@@ -8,9 +8,6 @@ from fastapi.security import OAuth2PasswordBearer
 import os
 
 # register user and hash password
-pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
-def hash_password(password: str) -> str:
-    return pwd_context.hash(password)
 
 def verify_password(plain_password: str, hashed_password: str) -> bool:
     return pwd_context.verify(plain_password, hashed_password)
