+++ db/requirements.txt
@@ -1,0 +1,3 @@
+pandas==2.2.3
+python-dotenv==1.1.0
+SQLAlchemy==2.0.41
