+++ docker-compose.yml
@@ -1,6 +1,6 @@
 services:
   db:
-    image: postgres
+    image: postgres:latest
     container_name: ${POSTGRES_DB}
     restart: always
     environment:
@@ -11,41 +11,70 @@ services:
       - ${LOCAL_PORT}:${POSTGRES_PORT}
     volumes:
       - pg_data:/var/lib/postgresql/data
-
-  wine-quality-db-init:
+    
+  db-init:
     build:
-      context: ..
-      dockerfile: ./docker/Dockerfile
-    container_name: wine-quality-db-init
+      context: .
+      dockerfile: db/Dockerfile
+    env_file:
+      - .env
     depends_on:
       - db
-    command: ["python", "init_db/init_db.py"]
-    volumes:
-      - .:/app
-    working_dir: /app
+    restart: "no"
 
-  wine-quality-api:
-    build:
-      context: ..
-      dockerfile: ./docker/Dockerfile
-    container_name: wine-quality-api
-    restart: always
+  mlflow:
+    image: bitnami/mlflow:latest
+    command: >
+      mlflow server
+      --backend-store-uri /mlflow
+      --default-artifact-root /mlflow
+      --host 0.0.0.0
     ports:
-      - "8000:80"
-    environment:
-      - POSTGRES_USER=${POSTGRES_USER}
-      - POSTGRES_PASSWORD=${POSTGRES_PASSWORD}
-      - POSTGRES_DB=${POSTGRES_DB}
-      - POSTGRES_HOST=${POSTGRES_HOST}
-      - POSTGRES_PORT=${POSTGRES_PORT}
-      - JWT_SECRET_KEY=${JWT_SECRET_KEY}
-      - JWT_ALGORITHM=${JWT_ALGORITHM}
-      - ACCESS_TOKEN_EXPIRE_MINUTES=${ACCESS_TOKEN_EXPIRE_MINUTES}
-      - TEST_USER=${TEST_USER}
-      - TEST_USER_PASSWORD=${TEST_USER_PASSWORD}
+      - ${MLFLOW_LOCAL_PORT}:${MLFLOW_PORT}
+    volumes:
+      - mlruns:/mlflow
     depends_on:
-      - wine-quality-db-init
+      - db
+      - db-init
+
+  training:
+    build:
+      context: .
+      dockerfile: training/Dockerfile
+    env_file:
+      - .env
+    volumes:
+      - mlruns:/app/mlruns
+      - model_volume:/app/default_model
+    depends_on:
+      - db
+      - db-init
+      - mlflow
+
+#   api:
+#     build:
+#       context: ..
+#       dockerfile: ./docker/Dockerfile
+#     container_name: wine-quality-api
+#     restart: always
+#     ports:
+#       - "8000:80"
+#     environment:
+#       - POSTGRES_USER=${POSTGRES_USER}
+#       - POSTGRES_PASSWORD=${POSTGRES_PASSWORD}
+#       - POSTGRES_DB=${POSTGRES_DB}
+#       - POSTGRES_HOST=${POSTGRES_HOST}
+#       - POSTGRES_PORT=${POSTGRES_PORT}
+#       - JWT_SECRET_KEY=${JWT_SECRET_KEY}
+#       - JWT_ALGORITHM=${JWT_ALGORITHM}
+#       - ACCESS_TOKEN_EXPIRE_MINUTES=${ACCESS_TOKEN_EXPIRE_MINUTES}
+#       - TEST_USER=${TEST_USER}
+#       - TEST_USER_PASSWORD=${TEST_USER_PASSWORD}
+#     depends_on:
+#       - wine-quality-db-init
 
 volumes:
   pg_data:
+  mlruns:
+  model_volume:
