--- docker/Dockerfile.train
@@ -1,15 +1,0 @@
-FROM python:3.11-slim
-
-RUN apt-get update && apt-get install -y --no-install-recommends gcc libpq-dev && rm -rf /var/lib/apt/lists/*
-
-WORKDIR /app
-
-COPY training/requirements.txt .
-RUN pip install --no-cache-dir -r requirements.txt
-
-COPY training/ .
-
-COPY ./docker/.env /app/.env
-
-CMD ["python", "train_model.py"]
-
