--- requirements.txt
@@ -1,59 +1,0 @@
-annotated-types==0.7.0
-anyio==4.9.0
-certifi==2025.4.26
-click==8.1.8
-contourpy==1.3.2
-cycler==0.12.1
-Deprecated==1.2.18
-dnspython==2.7.0
-dotenv==0.9.9
-ecdsa==0.19.1
-email_validator==2.2.0
-fastapi==0.115.12
-fonttools==4.57.0
-greenlet==3.2.1
-h11==0.16.0
-httpcore==1.0.9
-httpx==0.28.1
-idna==3.10
-iniconfig==2.1.0
-joblib==1.5.0
-jose==1.0.0
-kiwisolver==1.4.8
-limits==5.1.0
-matplotlib==3.10.1
-multipart==1.2.1
-numpy==2.2.5
-packaging==25.0
-pandas==2.2.3
-passlib==1.7.4
-pillow==11.2.1
-pluggy==1.5.0
-psycopg2-binary==2.9.10
-pyasn1==0.4.8
-pydantic==2.11.4
-pydantic_core==2.33.2
-pyparsing==3.2.3
-pytest==8.3.5
-pytest-asyncio==0.26.0
-python-dateutil==2.9.0.post0
-python-dotenv==1.1.0
-python-jose==3.4.0
-python-multipart==0.0.20
-pytz==2025.2
-PyYAML==6.0.2
-rsa==4.9.1
-scikit-learn==1.6.1
-scipy==1.15.2
-six==1.17.0
-slowapi==0.1.9
-sniffio==1.3.1
-SQLAlchemy==2.0.40
-starlette==0.46.2
-threadpoolctl==3.6.0
-tqdm==4.67.1
-typing-inspection==0.4.0
-typing_extensions==4.13.2
-tzdata==2025.2
-uvicorn==0.34.2
-wrapt==1.17.2
