+++ shared/requirements.txt
@@ -1,0 +1,4 @@
+passlib==1.7.4
+SQLAlchemy==2.0.41
+psycopg2-binary
+
