+++ training/requirements.txt
@@ -1,0 +1,5 @@
+click==8.2.1
+mlflow==2.22.0
+pandas==2.2.3
+python-dotenv==1.1.0
+scikit_learn==1.6.1
